+++ Filamented/FilamentCommonMaterial.cginc
@@ -12,49 +12,49 @@
 
 #define MIN_N_DOT_V 1e-4
 
-float clampNoV(float NoV) {
+half clampNoV(half NoV) {
     // Neubelt and Pettineo 2013, "Crafting a Next-gen Material Pipeline for The Order: 1886"
     return max(NoV, MIN_N_DOT_V);
 }
 
-float3 computeDiffuseColor(const float4 baseColor, float metallic) {
+half3 computeDiffuseColor(const half4 baseColor, half metallic) {
     return baseColor.rgb * (1.0 - metallic);
 }
 
-float3 computeF0(const float4 baseColor, float metallic, float reflectance) {
+half3 computeF0(const half4 baseColor, half metallic, half reflectance) {
     return baseColor.rgb * metallic + (reflectance * (1.0 - metallic));
 }
 
-float computeDielectricF0(float reflectance) {
+half computeDielectricF0(half reflectance) {
     return 0.16 * reflectance * reflectance;
 }
 
-float computeMetallicFromSpecularColor(const float3 specularColor) {
+half computeMetallicFromSpecularColor(const half3 specularColor) {
     return max3(specularColor);
 }
 
-float computeRoughnessFromGlossiness(float glossiness) {
+half computeRoughnessFromGlossiness(half glossiness) {
     return 1.0 - glossiness;
 }
 
-float perceptualRoughnessToRoughness(float perceptualRoughness) {
+half perceptualRoughnessToRoughness(half perceptualRoughness) {
     return perceptualRoughness * perceptualRoughness;
 }
 
-float roughnessToPerceptualRoughness(float roughness) {
+half roughnessToPerceptualRoughness(half roughness) {
     return sqrt(roughness);
 }
 
-float iorToF0(float transmittedIor, float incidentIor) {
+half iorToF0(half transmittedIor, half incidentIor) {
     return sq((transmittedIor - incidentIor) / (transmittedIor + incidentIor));
 }
 
-float f0ToIor(float f0) {
-    float r = sqrt(f0);
+half f0ToIor(half f0) {
+    half r = sqrt(f0);
     return (1.0 + r) / (1.0 - r);
 }
 
-float3 f0ClearCoatToSurface(const float3 f0) {
+half3 f0ClearCoatToSurface(const half3 f0) {
     // Approximation of iorTof0(f0ToIor(f0), 1.5)
     // This assumes that the clear coat layer has an IOR of 1.5
 #if FILAMENT_QUALITY == FILAMENT_QUALITY_LOW
