+++ Filamented/FilamentCommonOcclusion.cginc
@@ -13,18 +13,18 @@
 // Ambient occlusion helpers
 //------------------------------------------------------------------------------
 
-float SpecularAO_Lagarde(float NoV, float visibility, float roughness) {
+half SpecularAO_Lagarde(half NoV, half visibility, half roughness) {
     // Lagarde and de Rousiers 2014, "Moving Frostbite to PBR"
     return saturate(pow(NoV + visibility, exp2(-16.0 * roughness - 1.0)) - 1.0 + visibility);
 }
 
 #if defined(MATERIAL_HAS_BENT_NORMAL)
-float sphericalCapsIntersection(float cosCap1, float cosCap2, float cosDistance) {
+half sphericalCapsIntersection(half cosCap1, half cosCap2, half cosDistance) {
     // Oat and Sander 2007, "Ambient Aperture Lighting"
     // Approximation mentioned by Jimenez et al. 2016
-    float r1 = acosFastPositive(cosCap1);
-    float r2 = acosFastPositive(cosCap2);
-    float d  = acosFast(cosDistance);
+    half r1 = acosFastPositive(cosCap1);
+    half r2 = acosFastPositive(cosCap2);
+    half d  = acosFast(cosDistance);
 
     // We work with cosine angles, replace the original paper's use of
     // cos(min(r1, r2)) with max(cosCap1, cosCap2)
@@ -37,29 +37,29 @@ float sphericalCapsIntersection(float cosCap1, float cosCap2, float cosDistance)
         return 0.0;
     }
 
-    float delta = abs(r1 - r2);
-    float x = 1.0 - saturate((d - delta) / max(r1 + r2 - delta, 1e-4));
+    half delta = abs(r1 - r2);
+    half x = 1.0 - saturate((d - delta) / max(r1 + r2 - delta, 1e-4));
     // simplified smoothstep()
-    float area = sq(x) * (-2.0 * x + 3.0);
+    half area = sq(x) * (-2.0 * x + 3.0);
     return area * (1.0 - max(cosCap1, cosCap2));
 }
 #endif
 
 // This function could (should?) be implemented as a 3D LUT instead, but we need to save samplers
-float SpecularAO_Cones(float NoV, float visibility, float roughness) {
+half SpecularAO_Cones(half NoV, half visibility, half roughness) {
 #if defined(MATERIAL_HAS_BENT_NORMAL)
     // Jimenez et al. 2016, "Practical Realtime Strategies for Accurate Indirect Occlusion"
 
     // aperture from ambient occlusion
-    float cosAv = sqrt(1.0 - visibility);
+    half cosAv = sqrt(1.0 - visibility);
     // aperture from roughness, log(10) / log(2) = 3.321928
-    float cosAs = exp2(-3.321928 * sq(roughness));
+    half cosAs = exp2(-3.321928 * sq(roughness));
     // angle betwen bent normal and reflection direction
-    float cosB  = dot(shading_bentNormal, shading_reflected);
+    half cosB  = dot(shading_bentNormal, shading_reflected);
 
     // Remove the 2 * PI term from the denominator, it cancels out the same term from
     // sphericalCapsIntersection()
-    float ao = sphericalCapsIntersection(cosAv, cosAs, cosB) / (1.0 - cosAs);
+    half ao = sphericalCapsIntersection(cosAv, cosAs, cosB) / (1.0 - cosAs);
     // Smoothly kill specular AO when entering the perceptual roughness range [0.1..0.3]
     // Without this, specular AO can remove all reflections, which looks bad on metals
     return lerp(1.0, ao, smoothstep(0.01, 0.09, roughness));
@@ -71,7 +71,7 @@ float SpecularAO_Cones(float NoV, float visibility, float roughness) {
 /**
  * Computes a specular occlusion term from the ambient occlusion term.
  */
-float computeSpecularAO(float NoV, float visibility, float roughness) {
+half computeSpecularAO(half NoV, half visibility, half roughness) {
 #if SPECULAR_AMBIENT_OCCLUSION == SPECULAR_AO_SIMPLE
     return SpecularAO_Lagarde(NoV, visibility, roughness);
 #elif SPECULAR_AMBIENT_OCCLUSION == SPECULAR_AO_BENT_NORMALS
@@ -87,29 +87,29 @@ float computeSpecularAO(float NoV, float visibility, float roughness) {
  * The albedo term is meant to be the diffuse color or f0 for the diffuse and
  * specular terms respectively.
  */
-float3 gtaoMultiBounce(float visibility, const float3 albedo) {
+half3 gtaoMultiBounce(half visibility, const half3 albedo) {
     // Jimenez et al. 2016, "Practical Realtime Strategies for Accurate Indirect Occlusion"
-    float3 a =  2.0404 * albedo - 0.3324;
-    float3 b = -4.7951 * albedo + 0.6417;
-    float3 c =  2.7552 * albedo + 0.6903;
+    half3 a =  2.0404 * albedo - 0.3324;
+    half3 b = -4.7951 * albedo + 0.6417;
+    half3 c =  2.7552 * albedo + 0.6903;
 
     return max((visibility), ((visibility * a + b) * visibility + c) * visibility);
 }
 #endif
 
-void multiBounceAO(float visibility, const float3 albedo, inout float3 color) {
+void multiBounceAO(half visibility, const half3 albedo, inout half3 color) {
 #if MULTI_BOUNCE_AMBIENT_OCCLUSION == 1
     color *= gtaoMultiBounce(visibility, albedo);
 #endif
 }
 
-void multiBounceSpecularAO(float visibility, const float3 albedo, inout float3 color) {
+void multiBounceSpecularAO(half visibility, const half3 albedo, inout half3 color) {
 #if MULTI_BOUNCE_AMBIENT_OCCLUSION == 1 && SPECULAR_AMBIENT_OCCLUSION != SPECULAR_AO_OFF
     color *= gtaoMultiBounce(visibility, albedo);
 #endif
 }
 
-float singleBounceAO(float visibility) {
+half singleBounceAO(half visibility) {
 #if MULTI_BOUNCE_AMBIENT_OCCLUSION == 1
     return 1.0;
 #else
